+++ Assets/HandPoseVisualizer.cs
@@ -1,6 +1,6 @@
 /*
 *   Hand Pose
-*   Copyright (c) 2022 NatML Inc. All Rights Reserved.
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
 */
 
 namespace NatML.Visualizers {