+++ Packages/ai.natml.vision.hand-pose/Runtime/HandPosePredictor.cs
@@ -1,11 +1,12 @@
 /* 
 *   Hand Pose
-*   Copyright (c) 2022 NatML Inc. All Rights Reserved.
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
 */
 
 namespace NatML.Vision {
 
     using System;
+    using System.Threading.Tasks;
     using NatML.Features;
     using NatML.Internal;
     using NatML.Types;
@@ -18,24 +19,17 @@ namespace NatML.Vision {
 
         #region --Client API--
         /// <summary>
-        /// Create a hand pose predictor.
-        /// </summary>
-        /// <param name="model">Hand landmark ML model.</param>
-        public HandPosePredictor (MLModel model) => this.model = model as MLEdgeModel;
-
-        /// <summary>
         /// Detect hand pose in an image.
         /// </summary>
         /// <param name="inputs">Input image.</param>
         /// <returns>Detected hand.</returns>
         public Hand Predict (params MLFeature[] inputs) {
-            // Check
-            if (inputs.Length != 1)
-                throw new ArgumentException(@"Hand pose predictor expects a single feature", nameof(inputs));
-            // Check type
+            // Pre-process image
             var input = inputs[0];
-            if (!MLImageType.FromType(input.type))
-                throw new ArgumentException(@"Hand pose predictor expects an an array or image feature", nameof(inputs));  
+            if (input is MLImageFeature imageFeature) {
+                (imageFeature.mean, imageFeature.std) = model.normalization;
+                imageFeature.aspectMode = model.aspectMode;
+            }  
             // Predict
             var inputType = model.inputs[0] as MLImageType;
             using var inputFeature = (input as IMLEdgeFeature).Create(inputType);
@@ -51,13 +45,31 @@ namespace NatML.Vision {
             var result = new Hand(anchors, score, handedness, inputType.height);
             return result;
         }
+
+        /// <summary>
+        /// Dispose the model and release resources.
+        /// </summary>
+        public void Dispose () => model.Dispose();
+
+        /// <summary>
+        /// Create a hand pose predictor.
+        /// </summary>
+        /// <param name="configuration">Edge model configuration.</param>
+        /// <param name="accessKey">NatML access key.</param>
+        public static async Task<HandPosePredictor> Create (
+            MLEdgeModel.Configuration configuration = null,
+            string accessKey = null
+        ) {
+            var model = await MLEdgeModel.Create("@natsuite/hand-pose", configuration, accessKey);
+            var predictor = new HandPosePredictor(model);
+            return predictor;
+        }
         #endregion
 
 
         #region --Operations--
         private readonly MLEdgeModel model;
-
-        void IDisposable.Dispose () { } // Not used
+        private HandPosePredictor (MLModel model) => this.model = model as MLEdgeModel;
         #endregion
     }
 }