+++ Packages/ai.natml.vision.hand-pose/Changelog.md
@@ -1,2 +1,8 @@
+## 1.0.1
++ Added `HandPosePredictor.Create` static method for creating the hand pose predictor.
++ Removed `HandPosePredictor` public constructor. Use `HandPosePredictor.Create` method instead.
++ Upgraded to NatML 1.1.3.
++ Migrated sample to use VideoKit.
+
 ## 1.0.0
 + First release.